OpenClaw: Move proxy Go code to repo root as standard Go project layout
Move go.mod, go.sum, Makefile, cmd/, internal/, and docs/ from proxy/
to repo root. Rename module from proxy to srsx. Rename cmd/proxy-go to
cmd/proxy. Update all import paths and default static files path.
