+++ go.mod
@@ -1,0 +1,11 @@
+module srsx
+
+go 1.18
+
+require github.com/go-redis/redis/v8 v8.11.5
+
+require (
+	github.com/cespare/xxhash/v2 v2.1.2 // indirect
+	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
+	github.com/joho/godotenv v1.5.1 // indirect
+)
