+++ internal/errors/stack.go
