+++ openclaw/TOOLS.md
@@ -46,6 +46,10 @@ Skills are shared. Your setup is yours. Keeping them apart means you can update 
 - **Never `git push`** — William pushes himself
 - **Commit workflow:** `git diff --cached` → understand the changes → write title/description → `git commit -m "OpenClaw: ..."`
 - Title prefix: `OpenClaw:`
+- **Co-author for ACP Claude Code:** If Claude Code (ACP) was used to make the changes, add:
+  `Co-authored-by: Claude Opus 4.6 <noreply@anthropic.com>`
+- **Co-author for ACP Codex:** If Codex (ACP) was used to make the changes, add:
+  `Co-authored-by: chatgpt-codex-connector[bot] <199175422+chatgpt-codex-connector[bot]@users.noreply.github.com>`
 
 ### Go (GVM)
