--- proxy/.gitignore
@@ -1,10 +1,0 @@
-.idea
-.claude
-.vscode
-
-srs-proxy
-.go-formarted
-
-# For AI to ignore these files.
-.env
-docs/ignore-worklog.md
