--- proxy/Makefile
@@ -1,23 +1,0 @@
-.PHONY: all build test fmt clean run
-
-all: build
-
-build: fmt ./srs-proxy
-
-./srs-proxy: cmd/proxy-go/*.go internal/**/*.go
-	go build -o srs-proxy ./cmd/proxy-go
-
-test:
-	go test ./...
-
-fmt: ./.go-formarted
-
-./.go-formarted: cmd/proxy-go/*.go internal/**/*.go
-	touch .go-formarted
-	go fmt ./...
-
-clean:
-	rm -f srs-proxy .go-formarted
-
-run: fmt
-	go run ./cmd/proxy-go
